Introduce topic grid component
