+++ feature/topics/src/main/kotlin/uk/govuk/app/topics/ui/widget/TopicsWidget.kt
@@ -1,15 +1,10 @@
 package uk.govuk.app.topics.ui.widget
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.material3.TextButton
-import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -18,16 +13,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.window.core.layout.WindowWidthSizeClass
 import uk.govuk.app.design.ui.component.BodyBoldLabel
 import uk.govuk.app.design.ui.component.CompactButton
-import uk.govuk.app.design.ui.component.MediumVerticalSpacer
-import uk.govuk.app.design.ui.component.SmallHorizontalSpacer
 import uk.govuk.app.design.ui.component.Title3BoldLabel
 import uk.govuk.app.design.ui.theme.GovUkTheme
 import uk.govuk.app.topics.R
 import uk.govuk.app.topics.TopicsWidgetViewModel
 import uk.govuk.app.topics.ui.component.TopicVerticalCard
+import uk.govuk.app.topics.ui.component.TopicsGrid
 import uk.govuk.app.topics.ui.model.TopicItemUi
 
 @Composable
@@ -86,8 +79,14 @@ private fun TopicsWidgetContent(
 
         TopicsGrid(
             topics = topics,
-            onClick = onTopicClick
-        )
+        ) { modifier, topic ->
+            TopicVerticalCard(
+                icon = topic.icon,
+                title = topic.title,
+                onClick = { onTopicClick(topic.ref, topic.title) },
+                modifier = modifier
+            )
+        }
 
         if (displayShowAll) {
             val seeAllButtonText = stringResource(R.string.allTopicsButton)
@@ -98,87 +97,6 @@ private fun TopicsWidgetContent(
             )
         }
     }
-}
-
-@Composable
-private fun TopicsGrid(
-    topics: List<TopicItemUi>,
-    onClick: (String, String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val windowWidthSizeClass = currentWindowAdaptiveInfo().windowSizeClass.windowWidthSizeClass
-
-    // Todo - ideally this would be a lazy grid to gain from performance optimizations, however
-    //  nested lazy components are not allowed without a non-trivial workaround. The performance
-    //  impact should be negligible with the amount of items currently being displayed but we may
-    //  have to re-visit this in the future.
-    Column(modifier) {
-        val columnCount = getColumnCount(windowWidthSizeClass)
-        val rowCount = getRowCount(
-            topicsCount = topics.size,
-            columnCount = columnCount
-        )
-
-        for (rowIndex in 0 until rowCount) {
-            TopicsRow(
-                topics = topics,
-                columnCount = columnCount,
-                rowIndex = rowIndex,
-                onClick = onClick
-            )
-            MediumVerticalSpacer()
-        }
-    }
-}
-
-@Composable
-private fun TopicsRow(
-    topics: List<TopicItemUi>,
-    columnCount: Int,
-    rowIndex: Int,
-    onClick: (String, String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier.height(intrinsicSize = IntrinsicSize.Max)
-    ) {
-        for (columnIndex in 0 until columnCount) {
-            if (columnIndex > 0) {
-                SmallHorizontalSpacer()
-            }
-
-            val topicIndex = (rowIndex * columnCount) + columnIndex
-            if (topicIndex < topics.size) {
-                val topic = topics[topicIndex]
-                TopicVerticalCard(
-                    icon = topic.icon,
-                    title = topic.title,
-                    onClick = { onClick(topic.ref, topic.title) },
-                    modifier = Modifier
-                        .fillMaxHeight()
-                        .weight(1f)
-                )
-            } else {
-                Box(Modifier.weight(1f)) {  }
-            }
-
-            if (columnIndex < columnCount - 1) {
-                SmallHorizontalSpacer()
-            }
-        }
-    }
-}
-
-private fun getColumnCount(windowWidthSizeClass: WindowWidthSizeClass): Int {
-    return if (windowWidthSizeClass == WindowWidthSizeClass.COMPACT) 2 else 4
-}
-
-private fun getRowCount(topicsCount: Int, columnCount: Int): Int {
-    var rowCount = topicsCount / columnCount
-    if (topicsCount.mod(columnCount) > 0) {
-        rowCount += 1
-    }
-    return rowCount
 }
 
 @Preview(showBackground = true)
